Fix 3x slower major speed regression introduced in 98a9550260f7480d983a074812e600a8dfecd43f
The commit that introduced the problem was trying to optimise rendering
dense meshes with the opening decorator. However, it introduced more
calculation in the initial project load and stored more data on every
object. This resulted in models loading 3x slower for me which is pretty
major. The original problem needs further investigation.